config: move access log config to new directory
Prep for adding gRPC access log types.
