+++ source/server/config/network/BUILD
@@ -33,21 +33,6 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "file_access_log_lib",
-    srcs = ["file_access_log.cc"],
-    hdrs = ["file_access_log.h"],
-    external_deps = ["envoy_filter_network_http_connection_manager"],
-    deps = [
-        "//include/envoy/registry",
-        "//include/envoy/server:access_log_config_interface",
-        "//source/common/access_log:access_log_formatter_lib",
-        "//source/common/access_log:access_log_lib",
-        "//source/common/config:well_known_names",
-        "//source/common/protobuf",
-    ],
-)
-
-envoy_cc_library(
     name = "http_connection_manager_lib",
     srcs = ["http_connection_manager.cc"],
     hdrs = ["http_connection_manager.h"],
