+++ test/common/access_log/BUILD
@@ -42,7 +42,7 @@ envoy_cc_test(
         "//source/common/stats:stats_lib",
         "//source/common/upstream:upstream_includes",
         "//source/common/upstream:upstream_lib",
-        "//source/server/config/network:file_access_log_lib",
+        "//source/server/config/access_log:file_access_log_lib",
         "//test/common/upstream:utility_lib",
         "//test/mocks/access_log:access_log_mocks",
         "//test/mocks/event:event_mocks",
