+++ test/integration/BUILD
@@ -253,6 +253,7 @@ envoy_cc_test_library(
         "//source/server:hot_restart_nop_lib",
         "//source/server:server_lib",
         "//source/server:test_hooks_lib",
+        "//source/server/config/access_log:file_access_log_lib",
         "//source/server/config/http:buffer_lib",
         "//source/server/config/http:cors_lib",
         "//source/server/config/http:dynamo_lib",
@@ -267,7 +268,6 @@ envoy_cc_test_library(
         "//source/server/config/http:zipkin_lib",
         "//source/server/config/network:client_ssl_auth_lib",
         "//source/server/config/network:echo_lib",
-        "//source/server/config/network:file_access_log_lib",
         "//source/server/config/network:http_connection_manager_lib",
         "//source/server/config/network:mongo_proxy_lib",
         "//source/server/config/network:ratelimit_lib",
