+++ test/server/config/network/BUILD
@@ -15,8 +15,8 @@ envoy_cc_test(
         "//source/common/access_log:access_log_lib",
         "//source/common/config:well_known_names",
         "//source/common/dynamo:dynamo_filter_lib",
+        "//source/server/config/access_log:file_access_log_lib",
         "//source/server/config/network:client_ssl_auth_lib",
-        "//source/server/config/network:file_access_log_lib",
         "//source/server/config/network:http_connection_manager_lib",
         "//source/server/config/network:mongo_proxy_lib",
         "//source/server/config/network:ratelimit_lib",
