+++ test/server/config/network/config_test.cc
@@ -7,8 +7,8 @@
 #include "common/config/well_known_names.h"
 #include "common/dynamo/dynamo_filter.h"
 
+#include "server/config/access_log/file_access_log.h"
 #include "server/config/network/client_ssl_auth.h"
-#include "server/config/network/file_access_log.h"
 #include "server/config/network/http_connection_manager.h"
 #include "server/config/network/mongo_proxy.h"
 #include "server/config/network/ratelimit.h"
